requirements.txt
face_recognition
Pillow==7.2.0
tk
opencv-python
imutils